only extend top level main object with dsl methods
